Determine the age group
The focus of concepts are Constructors, Classes, and instances.
This program accepts age as an integer, passes it as an argument to the class constructor. With a set of constraints, the methods within the class perform the math and will display the age group with a statement. Try it yourself!

If you have a better approach, please write your code in the comments below. If I figure out a better solution in the following days, I will update the code.