A small scale phone directory
The focus of concepts is Dictionaries and instances.
This program n pair of names and numbers, I had to create a small scale phone directory to store the names and associated numbers. When quired for a name, the program should display the number. If the name does not exist, the program should show the appropriate message. The above function was built with certain logical constraints, takes in the arguments, and creates a dictionary with the key and value pair. When quired with a name, it responds accordingly. Try it yourself!

If you have a better approach, please write your code in the comments below. If I figure out a better solution in the following days, I will update the code.