Find the Binary value, count max consecutive 1s
The focus of concepts is building a robust algorithm that converts Base 10 integer to Base 2 value, counts the maximum number of consecutive 1's in the resulting array. 
This program accepts an integer, passes it as an argument to the first method where the value is converted. The logic differs if the given number is an even or an odd number. The resulting array from the first method passed as an argument to the second function, where the maximum number of consecutive 1s gets calculated. There are a few sets of constraints to be respected to deduce the algorithm. Try it yourself!