Find the height of a Binary Search Tree
The focus of concepts is to find the height of a Binary Search Tree. This program accepts an integer which defines the size of the array, passes it as an argument to the first method of second class where the left and right BST are constructed. It then returns the root node. The getHeight from the second class is called with root as an argument. getHeight is executed recursively to find the distance between the root, left subtree and right subtree. The max of either distance is returned.