Insert an integer at the tail of the Linked List
The focus of concepts is to perform insert operation for a Linked List
This program accepts age as an integer, passes it as an argument to the class constructor. With a set of constraints, the methods within the class will add data to the tail of the Linked list. Try it yourself!