Breadth First Search on BST
The focus of concepts is Level Order Traversal on a Binary search tree. This program accepts two inputs, one for the count of data and then the data itself. Insert method from Solution class builds the tree, I wrote the Queue class to facilitate the tree traversal. It hold helper functions to perform queue operations and returns the data to the levelOrder function from Solution class. Starting from the root, the algorithm checks if there are any child nodes, if the condition is true the data from left to right is added to the queue This step is repeated until the queue size is 0. The levelOrder finally prints the array of data in level order Try it yourself!