Remove duplicate nodes from the Linked List
This challenge made my day!
The focus of the concept is the Singly Linked List.
This program prompts for an integer value that sets the length of the linked list and prompts for an equal value number of data. My task was to complete the removeDuplicates method. I feel good about this solution because I tried to solve this without creating a temp variable; I managed to relocate the node reference without a temp variable. Try it yourself!
If you have a better approach, please write your code in the comments below. If I figure out a better solution in the following days, I will update the code.