Calculate the fine for late returns
The focus of the concept is Nested loop
This program accepts two lines of inputs. The first line is the actual date of return and the second line is expected date of return. The input is in DD MM YYYY format. Based on the input the fine is calculated. Try it yourself!

If you have a better approach, please write your code in the comments below. If I figure out a better solution in the following days, I will update the code.