Unit test the cases
The focus of the concept is Unit test
This program doesn't accept an explicit input but will generate arrays and return certain values when the functions are invoked. There are three classes each of these classes has at least two static methods and the last class has some test cases. My task was to return certain values from static methods to the test cases such that all the test cases are satisfied. Try it yourself!