I only love @gmail.com
The focus of the concept is the regular expression
This program accepts two inputs, one the integer which assigns the list size and the following lines accept the first name and email as input. The first name is appended to the list only if the email contains @gmail.com. This is then passed as an argument to sortInfo method. Sort info swill sort the names alphabetically and displays the name in a for loop. Try it yourself!

If you have a better approach, please write your code in the comments below. If I figure out a better solution in the following days, I will update the code.