Convert Romans numbers to Integers
In this solution, I have used the concept of dictionaries. Every roman alphabet is associated with a value, we receive the key and I have the value stored in the dictionary. It is easy that way to compare the values with the previous character and perform the necessary math operation. I love this problem. Try it yourself!