Check if the integer is a palindrome
After learning Time and space complexity, I am constantly motivated  to be a better programmer
This piece of code manages space well but is slightly low on time complexity performance. The program takes an integer input and the function will convert the integer to string, will create another string in reverse order. They both are compared to check for palindrome.