Sum the arrays in dynamic chunks
The program accepts two lines of input, the first one being the size of the array followed by the array elements.
I love the part how the length of the array is used as an advantage to get the desired output. I first tried with just m difference in the loop, the program did not go through a few of the test cases. This was excluding the last element in the array. Later I tried m-1 and got the right result.