Amongst the most frequent integers find the least
The challenge was to find the integer that has a maximum frequency if there is more than one integer with similar frequency the integer with the least value has to be returned. I think my code is efficient because the code is not designed only for integers between 1-5 but for any value and number of integers. Try it yourself!