Array addition, modulus comparison
The concepts involved in this program are array addition and modulus. I had to duplicate the given array for the sake of comparison and addition. The items from both the arrays are added and pared if it's modulus with the given number returns zero . The count variable is incremented when the condition is satisfied.