256th Julian or Georgian Day of the year.
The program accepts one four-digit integer input. The task is to find the 256th day considering the leap year. From reading the problem description it gets a little confusing because of the two calendar formats and leap year. It gets easy to have an array of days in months until September and then subtract the days in month array with 256. The remaining value will be the day of the year.