Check for palindrome, a three line solution
The program accepts a line of string, I used regEx and substring to reform the given string. In the second line of code, I compare the reformed string with the reverse of the reformed string. Note that I have used lower() to convert the string to lower case. Try this yourself!