Find number of integer pairs in an array
The story around the challenge was about a stock merchant. In the core of the problem, find the pair of integers and count the number of pairs in total discarding the odd pairs. My approach was to extract unique integers into another array, count the number of such integers from the unique array in parent array. Using // 2, add the values to get the number of pairs.