Treating int as str makes it easy
The task was to replace a digit in the given integer such that the resulting value should be the max value possible.
For a novice solver, the problem description can easily set a trap. Replacing 9 with 6 doesn't make any sense it will only reduce the value. Replacing the first occurrence of 6 with 9 will help get the max possible value. To avoid str to int and int to str operations I have used exponent multiplication to get the desired output. Try this yourself.