Dist b/w odd even, even odd, odd odd, even even
The task was to find the shortest path to reach a number with one or n as the starting point. On a number scale of 1 to 10, the digit 4 is closer to 1 as compared to 10. The problem description explained is associated with the pages of a book. There are two tricks One is the distance between start, end of the page and the page of interest. The other one is if the number of interest is odd, even and if the start point is odd, even. Based on these constraints, the if statement calculates the number of moves from the front and back to reach a page. Whichever is the least is displayed.