Find the number substrings with R=L
The task is to find the number of possible substrings with an equal number of R and L in it.
I used the for loop to create substrings and compare the R and L count in it. The runtime was quite low. Please let me know I can improve my code.