Find the second largest value in a nested list
The task was to find the second least score in a nested loop if there are more than one name associated with the score. Display the names in alphabetical order if there is more than one score with a similar second-largest score. I felt happy writing this piece of code. My solution to finding the second least score is better than other solutions. Also, with space complexity, it looks like I did a good job. Try it yourself!