Count number of times Down = Up
The task was to find the number of times there are Up will equal to the number od Down. I took the time to figure out the right answer despite the correct logic. The condition that checks for the equal number of down trail and equals the number of the up trails must have another condition; the unit has to be Up.