Defanging IP Address - replace . with [.]
The task was to find and replace '.' with '[.]'
Replace function; it only replaces the first occurrence unless the number of replacements is specified. My approach, loop through all the characters in the given string of valid IP Address, if the character is numeric, it will be added to the empty string I created else the character is replaced and then added to the empty string.