Count of negative integers in a 2D matrix
The task was to find the number of negative integers in a given 2D matrix. I followed the Brute force method. I di try to solve this by sorting the matrix, flatten the matrix using NumPy, enumerate function with the condition element < 0. But, the Brute force seemed to work best. Try it yourself!