Minimum Time Visiting All Points
The task was to find the shortest time required to travel from the start co-ordinates to the stop co-ordinates. Initially, I thought I could solve this with while loop and increment or decrement x, y value based on the next coordinates but I figured out an easy way with for loop. By taking the absolute difference between x, y coordinates of each point and add the max value returned with a time variable.