Sum of elements from two arrays
The program accepts three lines of input. The first line includes budget, number of keyboards, and drives available. The second line consists of the price of all the keyboards, and the last line accepts the price of drives. Using list comprehension, I summed the values from both the arrays and created another array that holds the sum that are less than or equal to the budget. Finally, the max value from the third array will be the result. If there are no elements in the third array, the program will return -1 as defined in the task.