Remove All Adjacent Duplicates in the string.
The task was to find the adjacent duplicate character and retain the rest from the string.
My approach was to find and remove any characters that have an even number of occurrences.  Although logically it works, it can not retain the order of characters. I got the above solution from another participant.